+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 **0.0.1-beta.1**
 
 * Initial commit
+
+**0.0.1-beta.2**
+
+* Bug fix for individual plugins not loading.
+* Added a new module to simplify package parsing.